requirements.txt
anyio==4.9.0
asgiref==3.9.1
certifi==2025.7.14
Django==5.2.4
djangorestframework==3.16.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
sniffio==1.3.1
sqlparse==0.5.3
typing_extensions==4.14.1
python-decouple==3.8
django-cors-headers==4.6.0
